+++ DependencyInjection/Compiler/SwaggerValidatorCompilerPass.php
@@ -13,10 +13,9 @@ declare(strict_types=1);
 
 namespace Linkin\Bundle\SwaggerResolverBundle\DependencyInjection\Compiler;
 
-use Linkin\Bundle\SwaggerResolverBundle\Factory\SwaggerResolverFactory;
+use Linkin\Bundle\SwaggerResolverBundle\Builder\SwaggerResolverBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -24,14 +23,20 @@ use Symfony\Component\DependencyInjection\Reference;
  */
 class SwaggerValidatorCompilerPass implements CompilerPassInterface
 {
+    public const TAG = 'linkin_swagger_resolver.validator';
+
     /**
      * {@inheritdoc}
      */
     public function process(ContainerBuilder $container): void
     {
+        if (!$container->hasDefinition(SwaggerResolverBuilder::class)) {
+            return;
+        }
+
         $validatorQueue = new \SplPriorityQueue();
 
-        foreach ($container->findTaggedServiceIds('linkin_swagger_resolver.validator') as $id => $attributes) {
+        foreach ($container->findTaggedServiceIds(self::TAG) as $id => $attributes) {
             $validatorReference = new Reference($id);
 
             $priority = isset($attributes['priority']) ? $attributes['priority'] : 0;
@@ -41,25 +46,9 @@ class SwaggerValidatorCompilerPass implements CompilerPassInterface
 
         $validators = iterator_to_array($validatorQueue);
 
-        $this->registerSwaggerResolverFactory($container, $validators);
-    }
-
-    /**
-     * @param ContainerBuilder $container
-     * @param array            $validators
-     */
-    private function registerSwaggerResolverFactory(ContainerBuilder $container, array $validators): void
-    {
-        $configurationLoaderId = $container->getParameter('linkin_swagger_resolver.configuration_loader');
-        $container->getParameterBag()->remove('linkin_swagger_resolver.configuration_loader');
-
-        $swaggerFactoryDefinition = new Definition(SwaggerResolverFactory::class);
-        $swaggerFactoryDefinition
-            ->addArgument($validators)
-            ->addArgument($container->getDefinition($configurationLoaderId))
+        $container
+            ->getDefinition(SwaggerResolverBuilder::class)
+            ->replaceArgument(0, $validators)
         ;
-
-        $container->setDefinition(SwaggerResolverFactory::class, $swaggerFactoryDefinition);
-        $container->setAlias('linkin_swagger_resolver.factory', SwaggerResolverFactory::class);
     }
 }